requirements.txt
Django
requests
Pillow
djangorestframework
django-utils-six
django-cors-headers
django-wysiwyg-redactor
mysqlclient
marshmallow